Changes of Data Type again

+++ database/migrations/2020_11_19_021613_create_couriers_table.php
@@ -31,7 +31,7 @@ class CreateCouriersTable extends Migration
             $table->string('recipient_pin');
             $table->string('description');
             $table->string('weight');
-            $table->bigint('price');
+            $table->integer('price');
             $table->string('status');
             $table->timestamps();
         });
